refactor(agent): extract conversation text helper
